Move all special keys handlers to single file
