+++ src/com/maddyhome/idea/vim/action/motion/leftright/MotionShiftHomeAction.kt
@@ -26,7 +26,7 @@ import com.maddyhome.idea.vim.command.Command
 import com.maddyhome.idea.vim.command.CommandFlags
 import com.maddyhome.idea.vim.command.MappingMode
 import com.maddyhome.idea.vim.group.MotionGroup
-import com.maddyhome.idea.vim.handler.specialkeys.ShiftedSpecialKeyHandler
+import com.maddyhome.idea.vim.handler.ShiftedSpecialKeyHandler
 import com.maddyhome.idea.vim.helper.vimForEachCaret
 import java.util.*
 import javax.swing.KeyStroke
